feat: add csrf_token where delete pin

+++ src/Controller/PinsController.php
@@ -79,10 +79,13 @@ class PinsController extends AbstractController
     /**
      * @Route("/pins/{id<[0-9]+>}/delete", name="app_pins_delete", methods={"DELETE"})
      */
-    public function delete(Pin $pin, EntityManagerInterface $em)
+    public function delete(Request $request, Pin $pin, EntityManagerInterface $em)
     {
-        $em->remove($pin);
-        $em->flush();
+        
+        if ($this->isCsrfTokenValid('pin_deletion_' . $pin->getId(), $request->request->get('csrf_token'))) {
+            $em->remove($pin);
+            $em->flush();
+        }
 
         return $this->redirectToRoute('app_home');
     }
